Refactor case persistence to reuse blob repository
